# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Doctoreto Wallet Voucher API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/transactions/callback": {
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Transactiom Callback",
                "description": "Transactiom Callback",
                "operationId": "transactionCallback",
                "requestBody": {
                    "description": "Pass Transaction info",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "transaction_id"
                                ],
                                "properties": {
                                    "transaction_id": {
                                        "type": "string",
                                        "example": "a0b53d8a-e9d1-4980-9e40-7370df4aa295"
                                    },
                                    "is_confirmed": {
                                        "type": "int",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transaction successfully confirmed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "code": {
                                                    "type": "int",
                                                    "example": 0
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "پرداخت شما با موفقیت تایید شد"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/api/v1/users/me/wallets": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get User Wallet",
                "description": "Get the Auth User Wallet Balance",
                "operationId": "getMyWallet",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "code": {
                                                    "type": "int",
                                                    "example": 0
                                                },
                                                "message": {
                                                    "properties": {
                                                        "balance": {
                                                            "type": "number",
                                                            "default": 0,
                                                            "example": 2000
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/api/v1/vouchers": {
            "post": {
                "tags": [
                    "Voucher"
                ],
                "summary": "Voucher Store",
                "description": "Make a single Voucher",
                "operationId": "storeVoucher",
                "requestBody": {
                    "description": "Pass voucher info",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "integer",
                                        "example": 2000
                                    },
                                    "usage_limit": {
                                        "type": "integer",
                                        "example": 1000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Voucher successfully created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "code": {
                                                    "type": "int",
                                                    "example": 0
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "EHngqXUWF"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/wallets/deposit": {
            "post": {
                "tags": [
                    "Wallet"
                ],
                "summary": "Wallet Deposit",
                "description": "Deposit Wallet Balance",
                "operationId": "walletDeposit",
                "requestBody": {
                    "description": "Pass Deposit info",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "integer",
                                        "example": 2000
                                    },
                                    "code": {
                                        "type": "string",
                                        "maxLength": 9,
                                        "example": "EHngqXUWF"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transaction successfully created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "code": {
                                                    "type": "int",
                                                    "example": 0
                                                },
                                                "message": {
                                                    "properties": {
                                                        "transaction_id": {
                                                            "type": "string",
                                                            "example": "fb7148bc-415e-4410-9e59-37ca5934f19f"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "Authorization": {
                "type": "apiKey",
                "description": "Fill with the User ID",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}